refactor(channel): extract spend methods to a separate class
